Separate out large test files in contract/
